+++ src/main/java/personal/william/raytracer/Sphere.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 public class Sphere implements SceneObject {
 
     private final Material material;
-    private final float radius;
+    private final double radius;
 
-    public Sphere(Material material, float radius) {
+    public Sphere(Material material, double radius) {
         this.material = material;
         this.radius = radius;
     }
@@ -21,13 +21,13 @@ public class Sphere implements SceneObject {
     public Optional<Vector3d> getFirstIntersection(
             Vector3d positioning, Vector3d orig, Vector3d dir) {
         Vector3d ray = positioning.minus(orig);
-        float rayToCenterDist = dir.dot(ray);
-        float centerDistSquare = ray.dot(ray)- (rayToCenterDist * rayToCenterDist);
-        float radiusSquare = radius * radius;
+        double rayToCenterDist = dir.dot(ray);
+        double centerDistSquare = ray.dot(ray)- (rayToCenterDist * rayToCenterDist);
+        double radiusSquare = radius * radius;
         if (centerDistSquare > (radiusSquare)) return Optional.empty();
 
         float intersectedDist = (float) Math.sqrt(radiusSquare - centerDistSquare);
-        float contactDist0 = rayToCenterDist - intersectedDist;
+        double contactDist0 = rayToCenterDist - intersectedDist;
         if (contactDist0 < 0) contactDist0 = rayToCenterDist + intersectedDist;
         return contactDist0 >= 0 ? Optional.of(dir.times(contactDist0)) : Optional.empty();
     }
